+++ app/src/app/permission/permission.run.js
@@ -6,7 +6,7 @@
         .run(permissionRun);
 
 
-    function permissionRun($rootScope, $cookies, $state, RoleStore, UserService) {
+    function permissionRun($rootScope, $cookies, $state, PermissionStore, RoleStore, UserService) {
         // normally this would be done at the login page but to show quick
         // demo we grab username from cookie and login the user
         var cookieUser = $cookies.get('tri-user');
@@ -14,15 +14,19 @@
             UserService.login(cookieUser);
         }
 
+        // create permissions and add check function verify all permissions
+        var permissions = ['viewEmail', 'viewGitHub', 'viewCalendar', 'viewLayouts', 'viewTodo', 'viewElements', 'viewAuthentication', 'viewCharts', 'viewMaps'];
+        PermissionStore.defineManyPermissions(permissions, function (permissionName) {
+            return UserService.hasPermission(permissionName);
+        });
+
         // create roles for app
-        RoleStore
-        .defineRole('SUPERADMIN', ['viewEmail', 'viewGitHub', 'viewCalendar', 'viewLayouts', 'viewTodo', 'viewElements', 'viewAuthentication', 'viewCharts', 'viewMaps'], checkRole);
-        RoleStore
-        .defineRole('ADMIN', ['viewLayouts', 'viewTodo', 'viewElements', 'viewAuthentication', 'viewCharts', 'viewMaps'], checkRole);
-        RoleStore
-        .defineRole('USER', ['viewAuthentication', 'viewCharts', 'viewMaps'], checkRole);
-        RoleStore
-        .defineRole('ANONYMOUS', [], checkRole);
+        RoleStore.defineManyRoles({
+            'SUPERADMIN': ['viewEmail', 'viewGitHub', 'viewCalendar', 'viewLayouts', 'viewTodo', 'viewElements', 'viewAuthentication', 'viewCharts', 'viewMaps'],
+            'ADMIN': ['viewLayouts', 'viewTodo', 'viewElements', 'viewAuthentication', 'viewCharts', 'viewMaps'],
+            'USER': ['viewAuthentication', 'viewCharts', 'viewMaps'],
+            'ANONYMOUS': []
+        });
 
 
         ///////////////////////
@@ -30,11 +34,6 @@
         // default redirect if access is denied
         function accessDenied() {
             $state.go('401');
-        }
-
-        // function that calls user service to check if permission is allowed
-        function checkRole(permission, transitionProperties) {
-            return UserService.hasPermission(permission, transitionProperties);
         }
 
         // watches